# web/openapi.json
{
  "components": {
    "schemas": {
      "Bar": {
        "properties": {
          "close": {
            "title": "Close",
            "type": "number"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "open": {
            "title": "Open",
            "type": "number"
          },
          "time": {
            "title": "Time",
            "type": "string"
          },
          "volume": {
            "title": "Volume",
            "type": "integer"
          }
        },
        "required": [
          "time",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "Bar",
        "type": "object"
      },
      "Brief": {
        "properties": {
          "conflicts": {
            "items": {
              "type": "string"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "generatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedat"
          },
          "notablePicks": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Notablepicks",
            "type": "array"
          },
          "overview": {
            "title": "Overview",
            "type": "string"
          },
          "pending": {
            "title": "Pending",
            "type": "boolean"
          },
          "positionNotes": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Positionnotes",
            "type": "array"
          },
          "state": {
            "enum": [
              "ready",
              "pending",
              "queued",
              "running",
              "skipped",
              "failed",
              "invalid_output"
            ],
            "title": "State",
            "type": "string"
          },
          "stateReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statereason"
          }
        },
        "required": [
          "date",
          "pending",
          "state",
          "stateReason",
          "generatedAt",
          "overview",
          "notablePicks",
          "conflicts",
          "positionNotes"
        ],
        "title": "Brief",
        "type": "object"
      },
      "BriefListItem": {
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "pending": {
            "title": "Pending",
            "type": "boolean"
          },
          "state": {
            "enum": [
              "ready",
              "pending",
              "queued",
              "running",
              "skipped",
              "failed",
              "invalid_output"
            ],
            "title": "State",
            "type": "string"
          },
          "stateReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statereason"
          }
        },
        "required": [
          "date",
          "pending",
          "state"
        ],
        "title": "BriefListItem",
        "type": "object"
      },
...
      "DecisionRequest": {
        "properties": {
          "confirm": {
            "default": false,
            "title": "Confirm",
            "type": "boolean"
          }
        },
        "title": "DecisionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreviewPick": {
        "description": "What a strategy the gate has NOT approved would pick. Never a recommendation.",
        "properties": {
          "company": {
            "title": "Company",
            "type": "string"
          },
          "exch": {
            "title": "Exch",
            "type": "string"
          },
          "holdDays": {
            "title": "Holddays",
            "type": "integer"
          },
          "horizon": {
            "title": "Horizon",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "number"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "signalDate": {
            "title": "Signaldate",
            "type": "string"
          },
          "stop": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop"
          },
          "strategy": {
            "title": "Strategy",
            "type": "string"
          },
          "strategyId": {
            "title": "Strategyid",
            "type": "string"
          },
          "strategyStatus": {
            "title": "Strategystatus",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "window": {
            "title": "Window",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "company",
          "exch",
          "horizon",
          "strategy",
          "strategyId",
          "strategyStatus",
          "score",
          "ref",
          "stop",
          "target",
          "window",
          "holdDays",
          "signalDate",
          "reason"
        ],
        "title": "PreviewPick",
        "type": "object"
      },
...
      "StockHit": {
        "properties": {
          "company": {
            "title": "Company",
            "type": "string"
          },
          "exch": {
            "title": "Exch",
            "type": "string"
          },
          "isin": {
            "title": "Isin",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "company",
          "exch",
          "isin"
        ],
        "title": "StockHit",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "stk",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/briefs": {
      "get": {
        "operationId": "briefs_api_briefs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BriefListItem"
                  },
                  "title": "Response Briefs Api Briefs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Briefs"
      }
    },
    "/api/briefs/{day}": {
      "get": {
        "operationId": "brief_api_briefs__day__get",
        "parameters": [
          {
            "in": "path",
            "name": "day",
            "required": true,
            "schema": {
              "title": "Day",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Brief"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Brief"
      }
    },
    "/api/briefs/{day}/generate": {
      "post": {
        "description": "Queue an evening review for a day.\n\nThe only thing this writes is a row in ``ai_requests``; ``stk ai worker`` does the work.\nThe API cannot call a model -- an import-linter contract keeps ``stk.ai`` out of\n``stk.api`` -- and this route is what makes that restriction bearable rather than a dead\nend. Idempotent: a day already queued or running is returned as-is, not queued twice.",
        "operationId": "generate_brief_api_briefs__day__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "day",
            "required": true,
            "schema": {
              "title": "Day",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Brief"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Brief"
      }
    },
...
    "/api/picks": {
      "get": {
        "operationId": "picks_api_picks_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Horizon"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Pick"
                  },
                  "title": "Response Picks Api Picks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Picks"
      }
    },
...
    "/api/previews": {
      "get": {
        "description": "What NOT-promoted strategies would pick. Informational: these are not recommendations.",
        "operationId": "previews_api_previews_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PreviewPick"
                  },
                  "title": "Response Previews Api Previews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Previews"
      }
    },
    "/api/proposals": {
      "get": {
        "operationId": "proposals_api_proposals_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalOut"
                  },
                  "title": "Response Proposals Api Proposals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proposals"
      }
    },
    "/api/proposals/{pid}/approve": {
      "post": {
        "description": "Approve a proposal. A NEW strategy must have passed the gate; a demotion needs confirm.",
        "operationId": "proposal_approve_api_proposals__pid__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DecisionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalOut"
                  },
                  "title": "Response Proposal Approve Api Proposals  Pid  Approve Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proposal Approve"
      }
    },
    "/api/proposals/{pid}/dismiss": {
      "post": {
        "operationId": "proposal_dismiss_api_proposals__pid__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalOut"
                  },
                  "title": "Response Proposal Dismiss Api Proposals  Pid  Dismiss Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proposal Dismiss"
      }
    },
    "/api/status": {
      "get": {
        "operationId": "status_api_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Status"
      }
    },
    "/api/stocks/search": {
      "get": {
        "operationId": "stock_search_api_stocks_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StockHit"
                  },
                  "title": "Response Stock Search Api Stocks Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stock Search"
      }
    },
    "/api/stocks/{symbol}": {
      "get": {
        "operationId": "stock_api_stocks__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stock"
      }
    },
    "/api/stocks/{symbol}/bars": {
      "get": {
        "operationId": "bars_api_stocks__symbol__bars_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Bar"
                  },
                  "title": "Response Bars Api Stocks  Symbol  Bars Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bars"
      }
    },
    "/api/strategies": {
      "get": {
        "operationId": "strategies_api_strategies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrategySummary"
                  },
                  "title": "Response Strategies Api Strategies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Strategies"
      }
    },
    "/api/strategies/{slug}": {
      "get": {
        "operationId": "strategy_api_strategies__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Strategy"
      }
    },
    "/api/strategies/{slug}/approve": {
      "post": {
        "description": "Take a gated candidate live. Refuses anything that has not passed the gate.",
        "operationId": "approve_api_strategies__slug__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve"
      }
    },
    "/api/strategies/{slug}/retire": {
      "post": {
        "description": "Take a strategy out of service. Requires explicit confirmation.",
        "operationId": "retire_api_strategies__slug__retire_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retire"
      }
    },
...
  }
}